[ADD] 포스터 선택 기능 구현
