+++ DOMO.swiftpm/Views/Poster/SelectPosterView.swift
@@ -6,44 +6,41 @@
 //
 
 import SwiftUI
+import NavigationStack
 
 struct SelectPosterView: View {
     var body: some View {
-        NavigationView {
-            ZStack {
-                Color.black
-                    .ignoresSafeArea()
+        ZStack {
+            Color.black
+                .ignoresSafeArea()
+            
+            VStack {
+                Text("당신의 선택은..?")
+                    .font(.custom(.DungGeunMo, size: 46))
+                    .foregroundColor(.white)
                 
-                VStack {
-                    Text("나의 매력을 어떻게 어필할 수 있을까?")
-                        .font(.custom(.DungGeunMo, size: 46))
-                        .foregroundColor(.white)
-                    
-                    Spacer()
-                        .frame(height: 58)
-                    
-                    VStack(spacing: 30) {
-                        ForEach(SelectionSample.allCases, id: \.self) { item in
-                            NavigationLink(destination: item.dstView) {
-                                item.textItem
-                                    .font(.custom(.gulim, size: 40))
-                                    .shadow(color: .black.opacity(0.25), radius: 1, x: 2, y: 2)
-                                    .foregroundColor(.black)
-                                    .frame(width: UIScreen.main.bounds.width * 0.83, height: UIScreen.main.bounds.height * 0.15)
-                                    .background(
-                                           LinearGradient(gradient: Gradient(
-                                            colors: [Color(red: 86 / 255, green: 199 / 255, blue: 224 / 255), Color(red: 221 / 255, green: 210 / 255, blue: 173 / 255)]),
-                                            startPoint: .top, endPoint: .bottom)
-                                       )
-                                    .cornerRadius(34)
-                            }
+                Spacer()
+                    .frame(height: 58)
+                
+                VStack(spacing: 30) {
+                    ForEach(SelectPosterList.allCases, id: \.self) { item in
+                        PushView(destination: item.dstView) {
+                            item.textItem
+                                .font(.custom(.gulim, size: 40))
+                                .shadow(color: .black.opacity(0.25), radius: 1, x: 2, y: 2)
+                                .foregroundColor(.black)
+                                .frame(width: UIScreen.main.bounds.width * 0.83, height: UIScreen.main.bounds.height * 0.15)
+                                .background(
+                                       LinearGradient(gradient: Gradient(
+                                        colors: [Color(red: 86 / 255, green: 199 / 255, blue: 224 / 255), Color(red: 221 / 255, green: 210 / 255, blue: 173 / 255)]),
+                                        startPoint: .top, endPoint: .bottom)
+                                   )
+                                .cornerRadius(34)
                         }
                     }
                 }
             }
         }
-        .navigationViewStyle(.stack)
-        .navigationBarBackButtonHidden(true)
         .ignoresSafeArea()
     }
 }
